+++ test/test.js
@@ -1,18 +1,13 @@
 import * as dom from '../Dom.js';
-import { LocalStorage } from "../Storage.js";
-import * as url from '../URL.js';
-
+import { toggle_component } from "../Components/Toggle.js";
 
 dom.on_page_load(() => {
-
-	const filters = {
-		category: ['fiction', 'history'],
-		sort: 'asc',
-		rating: { min: 4, max: 5 }
-	};
-	  
-	let string = url.to_string({value: "12312", a: 5});
-	let json = url.to_json(string);
-	console.log(json); 
-
+	let body = dom.get_body();
+	
+	const toggle_element = new toggle_component();
+	dom.elem("toggle").parent(body).text("Toggle").event("click", () => {
+		toggle_element.toggle();
+		console.log("Toggled: ", toggle_element.is_toggeled());
+	});
+	
 });