+++ changelog.md
@@ -2,6 +2,7 @@
 # 14-05-2025 :
 - added `Toggle.js` file in `Component` folder
 	- added `toggle_component` class
+- added `get_container_element, get_header_element, get_menu_element` methods to `dropdown_component` class
 
 # 13-05-2025 :
 - created a `Component` folder